# Optimize the performance of model.HashCode