**Please provide a description of this PR:**
issue #41603 

xxhash.Sum64 instead of md5.Sum for better performance.
benchmark：
```code
name                      old time/op    new time/op    delta
HashCode/small_string-4      381ns ±11%     115ns ±22%   -69.80%  (p=0.000 n=9+10)
HashCode/middle_string-4     488ns ±24%     204ns ±17%   -58.17%  (p=0.000 n=9+10)
HashCode/long_string-4       659ns ±13%     279ns ±11%   -57.71%  (p=0.000 n=10+8)

name                      old alloc/op   new alloc/op   delta
HashCode/small_string-4      16.0B ± 0%     16.0B ± 0%      ~     (all equal)
HashCode/middle_string-4     16.0B ± 0%    128.0B ± 0%  +700.00%  (p=0.000 n=10+10)
HashCode/long_string-4        160B ± 0%      288B ± 0%   +80.00%  (p=0.000 n=10+10)

name                      old allocs/op  new allocs/op  delta
HashCode/small_string-4       1.00 ± 0%      1.00 ± 0%      ~     (all equal)
HashCode/middle_string-4      1.00 ± 0%      2.00 ± 0%  +100.00%  (p=0.000 n=10+10)
HashCode/long_string-4        3.00 ± 0%      2.00 ± 0%   -33.33%  (p=0.000 n=10+10)
```