@hzxuzhonghu 2022-10-27
This increases memory allocations from benchmark