@istio-testing 2022-10-27
@wulianglongrd: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
gencheck_istio | b85dca040df5a563838bcfd87a6a250575dd4916 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/41642/gencheck_istio/1585542488095985664) | true | `/test gencheck`
lint_istio | b85dca040df5a563838bcfd87a6a250575dd4916 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/41642/lint_istio/1585542487261319168) | true | `/test lint`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->